+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "start": "webpack-dev-server --inline --progress --port 3000",
     "test": "karma start",
-    "build": "rimraf dist && webpack --config webpack.config.js --progress --profile --bail"
+    "build": "rimraf dist && webpack --config config/webpack.prod.js --progress --profile --bail"
   },
   "keywords": [
     "syncfusion",
@@ -27,14 +27,14 @@
   },
   "homepage": "https://github.com/syncfusion/angular2-seeds#readme",
   "dependencies": {
-    "@angular/common": "~5.0.0",
-    "@angular/compiler": "~5.0.0",
-    "@angular/core": "~5.0.0",
-    "@angular/forms": "~5.0.0",
-    "@angular/http": "~5.0.0",
-    "@angular/platform-browser": "~5.0.0",
-    "@angular/platform-browser-dynamic": "~5.0.0",
-    "@angular/router": "~5.0.0",
+    "@angular/common": "~5.1.2",
+    "@angular/compiler": "~5.1.2",
+    "@angular/core": "~5.1.2",
+    "@angular/forms": "~5.1.2",
+    "@angular/http": "~5.1.2",
+    "@angular/platform-browser": "~5.1.2",
+    "@angular/platform-browser-dynamic": "~5.1.2",
+    "@angular/router": "~5.1.2",
     "core-js": "^2.4.1",
     "rxjs": "^5.4.3",
     "zone.js": "^0.7.4",
