+++ webpack.config.js
@@ -1,199 +1,1 @@
-var helpers = require('./config/helpers');
-const fs = require('fs');
-const path = require('path');
-const ProgressPlugin = require('webpack/lib/ProgressPlugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const rxPaths = require('rxjs/_esm5/path-mapping');
-var webpack = require('webpack');
-const { NoEmitOnErrorsPlugin, SourceMapDevToolPlugin, NamedModulesPlugin } = require('webpack');
-const { CommonsChunkPlugin } = require('webpack').optimize;
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-const nodeModules = path.join(process.cwd(), 'node_modules');
-const realNodeModules = fs.realpathSync(nodeModules);
-const genDirNodeModules = path.join(process.cwd(), 'src', '$$_gendir', 'node_modules');
-const entryPoints = ["inline", "polyfills", "sw-register", "styles", "vendor", "main"];
-const ENV = process.env.NODE_ENV = process.env.ENV = 'production';
-module.exports = {
-    "resolve": {
-        "extensions": [
-            ".ts",
-            ".js"
-        ],
-        alias: {
-            jquery: "jquery/src/jquery"
-        },
-        "modules": [
-            "./node_modules",
-            "./node_modules"
-        ],
-        "symlinks": true,
-        "mainFields": [
-            "browser",
-            "module",
-            "main"
-        ]
-    },
-    "resolveLoader": {
-        "modules": [
-            "./node_modules",
-            "./node_modules"
-        ],
-        "alias": rxPaths()
-    },
-    "entry": {
-        "main": [
-            "./src\\main.ts"
-        ],
-        "polyfills": [
-            "./src\\polyfills.ts"
-        ],
-        "vendor": [
-            "./src\\vendor.ts"
-        ]
-    },
-    "output": {
-        "path": path.join(process.cwd(), "dist"),
-        "filename": "[name].bundle.js",
-        "chunkFilename": "[id].chunk.js",
-        "crossOriginLoading": false
-    },
-    "module": {
-        "rules": [
-            {
-                test: /\.ts$/,
-                loaders: [
-                    {
-                        loader: 'awesome-typescript-loader',
-                        options: { configFileName: helpers.root('src', 'tsconfig.json') }
-                    }, 'angular2-template-loader'
-                ]
-            },
-            {
-                "test": /\.html$/,
-                "loader": "raw-loader"
-            },
-            {
-                "test": /\.(eot|svg|cur)$/,
-                "loader": "file-loader",
-                "options": {
-                    "name": "[name].[hash:20].[ext]",
-                    "limit": 10000
-                }
-            },
-            {
-                "test": /\.(jpg|png|webp|gif|otf|ttf|woff|woff2|ani)$/,
-                "loader": "url-loader",
-                "options": {
-                    "name": "[name].[hash:20].[ext]",
-                    "limit": 10000
-                }
-            },
-            {
-                test: /\.css$/,
-                exclude: helpers.root('src', 'app'),
-                loader: ExtractTextPlugin.extract({ fallback: 'style-loader', loader: 'css-loader?sourceMap' })
-            },
-            {
-                test: /\.css$/,
-                include: helpers.root('src', 'app'),
-                loader: 'raw-loader'
-            }
-        ]
-    },
-    "plugins": [
-        new NoEmitOnErrorsPlugin(),
-        new ExtractTextPlugin('[name].css'),
-        new webpack.DefinePlugin({
-            'process.env': {
-                'ENV': JSON.stringify(ENV)
-            }
-        }),
-        new webpack.LoaderOptionsPlugin({
-            htmlLoader: {
-                minimize: false // workaround for ng2
-            }
-        }),
-        new HtmlWebpackPlugin({
-            "template": "./src\\index.html",
-            "filename": "./index.html",
-            "hash": false,
-            "inject": true,
-            "compile": true,
-            "favicon": false,
-            "minify": false,
-            "cache": true,
-            "showErrors": true,
-            "chunks": "all",
-            "excludeChunks": [],
-            "title": "Webpack App",
-            "xhtml": true,
-            "chunksSortMode": function sort(left, right) {
-                let leftIndex = entryPoints.indexOf(left.names[0]);
-                let rightindex = entryPoints.indexOf(right.names[0]);
-                if (leftIndex > rightindex) {
-                    return 1;
-                }
-                else if (leftIndex < rightindex) {
-                    return -1;
-                }
-                else {
-                    return 0;
-                }
-            }
-        }),
-        new CommonsChunkPlugin({
-            "name": [
-                "inline"
-            ],
-            "minChunks": null
-        }),
-        new CommonsChunkPlugin({
-            "name": [
-                "vendor"
-            ],
-            "minChunks": (module) => {
-                return module.resource
-                    && (module.resource.startsWith(nodeModules)
-                        || module.resource.startsWith(genDirNodeModules)
-                        || module.resource.startsWith(realNodeModules));
-            },
-            "chunks": [
-                "main"
-            ]
-        }),
-        new SourceMapDevToolPlugin({
-            "filename": "[file].map[query]",
-            "moduleFilenameTemplate": "[resource-path]",
-            "fallbackModuleFilenameTemplate": "[resource-path]?[hash]",
-            "sourceRoot": "webpack:///"
-        }),
-        new CommonsChunkPlugin({
-            "name": [
-                "main"
-            ],
-            "minChunks": 2,
-            "async": "common"
-        }),
-        new webpack.LoaderOptionsPlugin({
-            htmlLoader: {
-                minimize: false // workaround for ng2
-            }
-        }),
-        new NamedModulesPlugin({}),
-    ],
-    "node": {
-        "fs": "empty",
-        "global": true,
-        "crypto": "empty",
-        "tls": "empty",
-        "net": "empty",
-        "process": true,
-        "module": false,
-        "clearImmediate": false,
-        "setImmediate": false
-    },
-    "devServer": {
-        "historyApiFallback": true
-    }
-};
-
+module.exports = require('./config/webpack.dev.js');
